Refactor and enhance various components for improved stability and performance
- Updated `war_scene_download.c` to ensure proper type casting for UI text bounding dimensions.
- Enhanced error logging in `updateSceneDownload` to provide more informative messages.
- Added unused parameter macros in multiple state machine files to suppress compiler warnings.
- Changed sprite frame dimensions from `u8` to `u16` in `war_types.h` for better compatibility.
- Ensured proper type casting in sprite creation functions to prevent potential overflows.
- Improved pathfinding and entity management in state machine files for better performance.
- Updated UI handling in `war_ui.c` to ensure correct type casting for width settings.
- Renamed upgrade and command data arrays for clarity in `war_units.h`.
- Added assertions to validate dimensions in sprite creation to prevent invalid states.
- General code cleanup and consistency improvements across multiple files.

+++ src/war_state_machine_move.c
@@ -29,7 +29,7 @@ void enterMoveState(WarContext* context, WarEntity* entity, WarState* state)
     vec2 start = state->move.positions.items[state->move.positionIndex];
     vec2 end = state->move.positions.items[state->move.positionIndex + 1];
 
-    WarMapPath path = state->move.path = findPath(map->finder, start.x, start.y, end.x, end.y);
+    WarMapPath path = state->move.path = findPath(map->finder, (s32)start.x, (s32)start.y, (s32)end.x, (s32)end.y);
 
     // if the is no path to the next position, go to idle
     if (path.nodes.count <= 1)
@@ -44,10 +44,10 @@ void enterMoveState(WarContext* context, WarEntity* entity, WarState* state)
     }
 
     vec2 currentNode = path.nodes.items[state->move.pathNodeIndex];
-    setDynamicEntity(map->finder, currentNode.x, currentNode.y, unitSize.x, unitSize.y, entity->id);
+    setDynamicEntity(map->finder, (s32)currentNode.x, (s32)currentNode.y, (s32)unitSize.x, (s32)unitSize.y, entity->id);
 
     vec2 nextNode = path.nodes.items[state->move.pathNodeIndex + 1];
-    setDynamicEntity(map->finder, nextNode.x, nextNode.y, unitSize.x, unitSize.y, entity->id);
+    setDynamicEntity(map->finder, (s32)nextNode.x, (s32)nextNode.y, (s32)unitSize.x, (s32)unitSize.y, entity->id);
 
     setUnitDirectionFromDiff(entity, nextNode.x - currentNode.x, nextNode.y - currentNode.y);
     setAction(context, entity, WAR_ACTION_TYPE_WALK, false, getUnitActionScale(entity));
@@ -214,7 +214,7 @@ void updateMoveState(WarContext* context, WarEntity* entity, WarState* state)
         }
 
         currentNode = path->nodes.items[state->move.pathNodeIndex];
-        setDynamicEntity(map->finder, currentNode.x, currentNode.y, unitSize.x, unitSize.y, entity->id);
+        setDynamicEntity(map->finder, (s32)currentNode.x, (s32)currentNode.y, (s32)unitSize.x, (s32)unitSize.y, entity->id);
 
         nextNode = path->nodes.items[state->move.pathNodeIndex + 1];
 
@@ -242,7 +242,7 @@ void updateMoveState(WarContext* context, WarEntity* entity, WarState* state)
             return;
         }
 
-        setDynamicEntity(map->finder, nextNode.x, nextNode.y, unitSize.x, unitSize.y, entity->id);
+        setDynamicEntity(map->finder, (s32)nextNode.x, (s32)nextNode.y, (s32)unitSize.x, (s32)unitSize.y, entity->id);
         setUnitDirectionFromDiff(entity, nextNode.x - currentNode.x, nextNode.y - currentNode.y);
     }
 }
